fix: repair12 docker

+++ gateway/src/main/java/ru/practicum/shareit/item/dto/CommentDto.java
@@ -14,7 +14,7 @@ public class CommentDto {
     private final Long id;
     @NotEmpty(message = "Text cannot be null")
     private final String text;
-    private final Integer itemId;
+    private final Long itemId;
     private final String authorName;
     private final LocalDateTime created;
 }
